- change swap deploy script

+++ scripts/deploy_swap_file.sh
@@ -8,10 +8,13 @@
 # https://wiki.archlinux.org/index.php/swap
 #
 
+if ! free | awk '/^Swap:/ {exit !$2}';
+then
 
-sudo dd if=/dev/zero of=/swapspace bs=1M count=4000
+    sudo dd if=/dev/zero of=/swapspace bs=1M count=4000
 
-sudo mkswap /swapspace
-sudo swapon /swapspace
+    sudo mkswap /swapspace
+    sudo swapon /swapspace
 
-echo 'swap file created'
+    echo 'swap file created'
+fi
